Docs are regenerated, memory-api is updated, license terms are added to each source/header file

+++ src/interpreter.c
@@ -1,3 +1,17 @@
+// This file is part of fbgc
+
+// fbgc is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// fbgc is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with fbgc.  If not, see <https://www.gnu.org/licenses/>.
 #include "fbgc.h"
 
 
@@ -528,7 +542,7 @@ struct fbgc_object * run_code(struct interpreter_packet * ip){
 
 			if(funo->type == CFUN){
 				STACK_GOTO(-arg_no - kwargs_flag);
-				struct fbgc_cfun_arg cfarg = {.arg =sp+sctr, .argc = arg_no, .kwargs_flag = kwargs_flag };
+				struct fbgc_cfun_arg cfarg = {.arg =sp+sctr, .argc = (uint8_t)arg_no, .kwargs_flag = kwargs_flag };
 				struct fbgc_object * res = cast_fbgc_object_as_cfun(funo)->function(&cfarg);
 				
 				--pop_number;
